+++ changelog.md
@@ -2,6 +2,7 @@
 
 - macro input
     - `map MapType` is not followed by `with iter IterType` anymore
+    - `range RangeType` has been removed, use `..` and `..=` operators instead
 - maps:
     - no dedicated iterator type anymore
     - function `of` constructing a map from a vector is gone, use the `From` implementation
@@ -10,6 +11,7 @@
     - `last_mut` has been fixed and also returns the last element's index;
     - new `push_idx` function: like `push`, but takes an element constructor taking the element's
         index as argument;
+    - `split` now produces iterators that yield indices
     - added implementation of `std::ops::Index<std::ops::RangeToInclusive<usize>>`
 
 # v0.9.11